requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.39.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scipy==1.15.3
seaborn==0.13.2
setuptools==80.4.0
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.133
ultralytics-thop==2.0.14
urllib3==2.4.0
watchdog==6.0.0
